Get all classical RNG measures

Get all classical RNG measures

allRNG(
  y,
  minScale = NA,
  maxScale = NA,
  responseAlternatives = NA,
  results = c("classical", "randseqR")[2],
  Redundancy = TRUE,
  RNG = TRUE,
  RNG2 = TRUE,
  RF = TRUE,
  Coupon = TRUE,
  NSQ = TRUE,
  FOD = TRUE,
  Adjacency = TRUE,
  TPI = TRUE,
  PhL = TRUE,
  Runs = TRUE,
  repDistance = TRUE,
  repGap = TRUE,
  PhiIndex = TRUE
)

Arguments

y
A sequence of symbols. If y is non-numeric, unique elements will be labelled by an integer value.

minScale
Minimum expected value. If y is a character vector this should refer to the lowest numeric code used.

maxScale
Maximum expected value. If y is a character vector this should refer to the highest numeric code used.

responseAlternatives
An optional vector of possible response alternatives. If NA, responseAlternatives will be set to seq(minScale,maxScale)

results
Either 'randseqR' (default) or 'classical'.

RNG
RNG (default = TRUE)

RNG2
RNG2 (default = TRUE)

RF
Response frequencies (default = TRUE)

Coupon
Coupon (default = TRUE)

NSQ
NSQ (default = TRUE)

Adjacency
Adjacency (default = TRUE)

TPI
TPI (default = TRUE)

Runs
Runs (default = TRUE)

R
Redundancy (default = TRUE)

PhaseLength
PhaseLength (default = TRUE)

RepDistance
Repetition Distance (default = TRUE)

RepGap
Repetition Gap (default = TRUE)

Phi
Phi index (default = TRUE)

Value

A list object of classical RNG task measures.

Examples

y <- round(runif(100,1,9))

# Omit the Frequency tables
allRNG(y, minScale = 1, maxScale = 9, RF = FALSE)

#> $Results
#> [1] "randseqR"
#> 
#> $N
#> [1] 100
#> 
#> $Redundancy
#> [1] 2.411093
#> attr(,"Name")
#> [1] "Redundancy"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> 
#> $RNG
#> [1] 27.99267
#> attr(,"Name")
#> [1] "RNG"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> attr(,"results")
#> [1] "randseqR"
#> 
#> $RNG2
#> [1] 27.78005
#> attr(,"Name")
#> [1] "RNG2"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> attr(,"results")
#> [1] "randseqR"
#> 
#> $RF
#> [1] FALSE
#> 
#> $Coupon
#> [1] 31.66667
#> attr(,"Name")
#> [1] "Coupon"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> 
#> $NSQ
#> [1] 31.25
#> attr(,"Name")
#> [1] "NSQ"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> attr(,"results")
#> [1] "randseqR"
#> 
#> $FOD
#> $FOD$`Diff_-7`
#> [1] 3
#> 
#> $FOD$`Diff_-6`
#> [1] 4
#> 
#> $FOD$`Diff_-5`
#> [1] 4
#> 
#> $FOD$`Diff_-4`
#> [1] 8
#> 
#> $FOD$`Diff_-3`
#> [1] 4
#> 
#> $FOD$`Diff_-2`
#> [1] 7
#> 
#> $FOD$`Diff_-1`
#> [1] 11
#> 
#> $FOD$Diff_0
#> [1] 11
#> 
#> $FOD$Diff_1
#> [1] 9
#> 
#> $FOD$Diff_2
#> [1] 11
#> 
#> $FOD$Diff_3
#> [1] 14
#> 
#> $FOD$Diff_4
#> [1] 7
#> 
#> $FOD$Diff_5
#> [1] 2
#> 
#> $FOD$Diff_6
#> [1] 3
#> 
#> $FOD$Diff_8
#> [1] 1
#> 
#> attr(,"Name")
#> [1] "FOD"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> attr(,"results")
#> [1] "randseqR"
#> 
#> $Adjacency
#> $Adjacency$Ascending
#> [1] 9
#> 
#> $Adjacency$Descending
#> [1] 11
#> 
#> $Adjacency$Combined
#> [1] 20
#> 
#> 
#> $TPI
#> [1] 91.83673
#> attr(,"Name")
#> [1] "TPI"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> 
#> $phaselength
#> $phaselength$PhL_1
#> [1] 30
#> 
#> $phaselength$PhL_2
#> [1] 21
#> 
#> $phaselength$PhL_3
#> [1] 7
#> 
#> $phaselength$PhL_4
#> [1] 1
#> 
#> attr(,"Name")
#> [1] "PhL"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> 
#> $Runs
#> [1] 0.5780245
#> attr(,"Name")
#> [1] "Runs"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> attr(,"results")
#> [1] "randseqR"
#> 
#> $repDistance
#> $repDistance$RD_1
#> [1] 11
#> 
#> $repDistance$RD_2
#> [1] 8
#> 
#> $repDistance$RD_3
#> [1] 11
#> 
#> $repDistance$RD_4
#> [1] 9
#> 
#> $repDistance$RD_5
#> [1] 3
#> 
#> $repDistance$RD_6
#> [1] 3
#> 
#> $repDistance$RD_7
#> [1] 8
#> 
#> $repDistance$RD_8
#> [1] 10
#> 
#> $repDistance$RD_9
#> [1] 5
#> 
#> $repDistance$RD_10
#> [1] 1
#> 
#> $repDistance$RD_12
#> [1] 1
#> 
#> $repDistance$RD_13
#> [1] 4
#> 
#> $repDistance$RD_14
#> [1] 3
#> 
#> $repDistance$RD_17
#> [1] 3
#> 
#> $repDistance$RD_18
#> [1] 2
#> 
#> $repDistance$RD_19
#> [1] 2
#> 
#> $repDistance$RD_21
#> [1] 2
#> 
#> $repDistance$RD_24
#> [1] 1
#> 
#> $repDistance$RD_33
#> [1] 1
#> 
#> $repDistance$RD_37
#> [1] 1
#> 
#> $repDistance$RD_42
#> [1] 1
#> 
#> $repDistance$RD_50
#> [1] 1
#> 
#> attr(,"Name")
#> [1] "repetition Distance"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#> 
#> $repGap
#> $repGap$RG_median
#> [1] 7
#> 
#> $repGap$RG_mean
#> [1] 8.56044
#> 
#> $repGap$RG_mode
#> [1] 1
#> 
#> 
#> $PhiIndex
#> $PhiIndex$Phi_2
#> [1] -0.4771527
#> 
#> $PhiIndex$Phi_3
#> [1] -1.801841
#> 
#> $PhiIndex$Phi_4
#> [1] 0.7621922
#> 
#> $PhiIndex$Phi_5
#> [1] -0.4950468
#> 
#> $PhiIndex$Phi_6
#> [1] -1.458568
#> 
#> $PhiIndex$Phi_7
#> [1] -2.502464
#> 
#> attr(,"Name")
#> [1] "phiIndex"
#> attr(,"y")
#>   [1] 2 8 6 2 1 5 5 3 7 7 8 2 1 4 4 3 4 2 4 9 3 6 7 3 9 7 1 5 7 7 1 3 3 6 5 4 7
#>  [38] 9 2 3 6 5 6 6 2 7 7 9 9 4 5 4 2 5 5 7 3 7 2 4 8 3 6 4 6 3 9 5 5 3 5 4 1 5
#>  [75] 4 1 4 5 8 4 5 6 3 3 5 8 4 3 6 1 9 2 5 8 7 3 5 8 1 5
#> attr(,"y")attr(,"transformation")
#> [1] "none"
#> attr(,"y")attr(,"minScale")
#> [1] 1
#> attr(,"y")attr(,"maxScale")
#> [1] 9
#> attr(,"y")attr(,"responseAlternatives")
#> [1] 1 2 3 4 5 6 7 8 9
#>