Get all classical RNG measures

Get all classical RNG measures

allRNG(y, minScale = NA, maxScale = NA, responseAlternatives = NA,
  R = TRUE, RNG = TRUE, RNG2 = TRUE, RF = TRUE, Coupon = TRUE,
  NSQ = TRUE, Adjacency = TRUE, TPI = TRUE, PhaseLength = TRUE,
  Runs = TRUE, RepDistance = TRUE, RepGap = TRUE, Phi = TRUE)

Arguments

y
A sequence of symbols. If y is non-numeric, unique elements will be labelled by an integer value.

minScale
Minimum expected value. If y is a character vector this should refer to the lowest numeric code used.

maxScale
Maximum expected value. If y is a character vector this should refer to the highest numeric code used.

responseAlternatives
An optional vector of possible response alternatives. If NA, responseAlternatives will be set to seq(minScale,maxScale)

R
Redundancy (default = TRUE)

RNG
RNG (default = TRUE)

RNG2
RNG2 (default = TRUE)

RF
Response frequencies (default = TRUE)

Coupon
Coupon (default = TRUE)

NSQ
NSQ (default = TRUE)

Adjacency
Adjacency (default = TRUE)

TPI
TPI (default = TRUE)

PhaseLength
PhaseLength (default = TRUE)

Runs
Runs (default = TRUE)

RepDistance
Repetition Distance (default = TRUE)

RepGap
Repetition Gap (default = TRUE)

Phi
Phi index (default = TRUE)

Value

A list object of classical RNG task measures.

Examples

y <- round(runif(100,1,9))

# Omit the Frequecny tables
allRNG(y, minScale = 1, maxScale = 9, RF = FALSE)

#> Error in check_y(y, minScale = minScale, maxScale = maxScale, responseAlternatives = responseAlternatives): Elements in y do not correspond to elements of responseAlternatives!